requirements.txt
keras==3.7.0
matplotlib==3.10.0
mlflow==2.19.0
mlflow_skinny==2.19.0
numpy==2.2.0
pandas==2.2.3
pyngrok==7.2.2
scikit_learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
tensorflow_intel==2.18.0
